requirements.txt
groq
pandas
python-dotenv==1.0.1